+++ Sandboxie/apps/control/MyFrame.h
@@ -127,6 +127,8 @@ class CMyFrame : public CFrameWnd
     afx_msg void OnCmdHelpUpdate();
     afx_msg void OnCmdHelpUpgrade();
     afx_msg void OnCmdHelpMigrate();
+    afx_msg void OnCmdHelpGetCert();
+    afx_msg void OnCmdHelpSetCert();
     afx_msg void OnCmdHelpAbout();
 	//afx_msg LRESULT OnUpdateResult(WPARAM wParam, LPARAM lParam);
     afx_msg void OnCmdTerminateProcess();