+++ Sandboxie/core/drv/verify.c
@@ -487,12 +487,13 @@ union _SCertInfo {
     ULONGLONG	State;
     struct {
         ULONG
-            valid     : 1, // certificate is active
-            expired   : 1, // certificate is expired but may be active
-            outdated  : 1, // certificate is expired, not anymore valid for the current build
-            business  : 1, // certificate is suitable for business use
-            evaluation: 1, // evaluation certificate
-            reservd_1 : 3,
+            valid     : 1,      // certificate is active
+            expired   : 1,      // certificate is expired but may be active
+            outdated  : 1,      // certificate is expired, not anymore valid for the current build
+            business  : 1,      // certificate is suitable for business use
+            evaluation: 1,      // evaluation certificate
+            grace_period: 1,    // the certificate is expired and or outdated but we keep it valid for 1 extra month to allof wor a seamless renewal
+            reservd_1 : 2,
             reservd_2 : 8,
             reservd_3 : 8,
             reservd_4 : 8;
@@ -728,21 +729,34 @@ _FX NTSTATUS KphValidateCertificate(void)
             } \
             Verify_CertInfo.expirers_in_sec = (ULONG)(((cert_date.QuadPart + KphGetDateInterval(days, months, years)) - LocalTime.QuadPart) / 10000000ll); // 100ns steps -> 1sec
 
+        // certs with a validity >= 3 months get 1 extra month of functionality
+#define TEST_GRACE_PERIODE(days, months, years) \
+                if (months >= 3 || years > 0){ \
+                    if ((cert_date.QuadPart + KphGetDateInterval(days, months + 1, years)) >= LocalTime.QuadPart) \
+                        Verify_CertInfo.grace_period = 1; \
+                } \
+
         // Check if the certificate is valid for the current build, failing this locks features out
 #define TEST_VALIDITY(days, months, years) \
             TEST_CERT_DATE(days, months, years) \
             if ((cert_date.QuadPart + KphGetDateInterval(days, months, years)) < BuildDate.QuadPart){ \
                 Verify_CertInfo.outdated = 1; \
-                Verify_CertInfo.valid = 0; \
-                status = STATUS_ACCOUNT_EXPIRED; \
+                TEST_GRACE_PERIODE(days, months, years) \
+                if(!Verify_CertInfo.grace_period){ \
+                    Verify_CertInfo.valid = 0; \
+                    status = STATUS_ACCOUNT_EXPIRED; \
+                } \
             }
 
         // Check if the certificate is expired, failing this locks features out
 #define TEST_EXPIRATION(days, months, years) \
             TEST_CERT_DATE(days, months, years) \
             if(Verify_CertInfo.expired == 1) { \
-                Verify_CertInfo.valid = 0; \
-                status = STATUS_ACCOUNT_EXPIRED; \
+                TEST_GRACE_PERIODE(days, months, years) \
+                if(!Verify_CertInfo.grace_period){ \
+                    Verify_CertInfo.valid = 0; \
+                    status = STATUS_ACCOUNT_EXPIRED; \
+                } \
             }
 
 
@@ -774,7 +788,7 @@ _FX NTSTATUS KphValidateCertificate(void)
                 // 
             } 
             else if (level && _wcsicmp(level, L"LARGE") == 0 && cert_date.QuadPart < KphGetDate(1,04,2022)) { // valid for all builds released with 2 years
-                TEST_CERT_DATE(0, 0, 2); // no real expiration just ui reminder
+                TEST_CERT_DATE(0, 0, 2); // no real expiration just ui reminder - old certs
             }
             else if (level && _wcsicmp(level, L"LARGE") == 0) { // valid for all builds released with 2 years
                 TEST_VALIDITY(0, 0, 2);