+++ SandboxiePlus/SandMan/Windows/SettingsWindow.h
@@ -171,12 +171,13 @@ union SCertInfo {
     quint64	State;
     struct {
         quint32
-            valid     : 1, // certificate is active
-            expired   : 1, // certificate is expired but may be active
-            outdated  : 1, // certificate is expired, not anymore valid for the current build
-            business  : 1, // certificate is suitable for business use
-            evaluation: 1, // evaluation certificate
-            reservd_1 : 3,
+            valid     : 1,      // certificate is active
+            expired   : 1,      // certificate is expired but may be active
+            outdated  : 1,      // certificate is expired, not anymore valid for the current build
+            business  : 1,      // certificate is suitable for business use
+            evaluation: 1,      // evaluation certificate
+            grace_period: 1,    // the certificate is expired and or outdated but we keep it valid for 1 extra month to allof wor a seamless renewal
+            reservd_1 : 2,
             reservd_2 : 8,
             reservd_3 : 8,
 			reservd_4 : 8;